[BE] feat(character): 턴 결과와 커리어·플레이 이력 규칙을 반영한다 (S14P21C103-196)
- 턴 결과 반영 시 스탯, 커리어, 플레이 이력 저장 흐름을 정리한다
- 번아웃, 강제 퇴사, 실업 급여 처리 규칙을 도메인과 서비스에 반영한다
- gameplay history 저장 필드와 조회 순서를 정리하고 관련 테스트를 보강한다

+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/domain/history/news/GameNewsLogRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GameNewsLogRepository extends JpaRepository<GameNewsLog, Integer> {
 
-    List<GameNewsLog> findAllByGameSessionIdOrderByTurnNumberDescGameNewsLogIdDesc(Integer gameSessionId);
+    List<GameNewsLog> findAllByGameSessionIdOrderByTurnNumberDescGameNewsLogIdDesc(Long gameSessionId);
 }
